# openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "servers": [{"url": "/"}], "paths": {"/api/v1/stashes/": {"get": {"summary": "List Stashes", "operationId": "list_stashes_api_v1_stashes__get", "parameters": [{"name": "page", "in": "query", "required": true, "schema": {"type": "integer", "exclusiveMinimum": 0, "title": "Page"}}, {"name": "take", "in": "query", "required": true, "schema": {"type": "integer", "exclusiveMaximum": 200, "exclusiveMinimum": 0, "title": "Take"}}, {"name": "show_revoked", "in": "query", "required": false, "schema": {"type": "boolean", "default": false, "title": "Show Revoked"}}, {"name": "show_expired", "in": "query", "required": false, "schema": {"type": "boolean", "default": false, "title": "Show Expired"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/ListStashesRow"}, "title": "Response List Stashes Api V1 Stashes  Get"}}}}, "500": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Internal Server Error"}, "403": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Forbidden"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/stashes/{slug}": {"delete": {"summary": "Revoke Stash", "operationId": "revoke_stash_api_v1_stashes__slug__delete", "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string", "title": "Slug"}}], "responses": {"204": {"description": "Successful Response"}, "500": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Internal Server Error"}, "403": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Forbidden"}, "404": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Not Found"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/stashes/text": {"post": {"summary": "Add Text Stash", "operationId": "add_text_stash_api_v1_stashes_text_post", "parameters": [{"name": "expires_at", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Expires At"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_add_text_stash_api_v1_stashes_text_post"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Stash"}}}}, "500": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Internal Server Error"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/stashes/file": {"post": {"summary": "Add Binary Stash", "operationId": "add_binary_stash_api_v1_stashes_file_post", "parameters": [{"name": "expires_at", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Expires At"}}], "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_add_binary_stash_api_v1_stashes_file_post"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Stash"}}}}, "500": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Internal Server Error"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Unprocessable Content"}, "413": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Content Too Large"}}}}, "/api/v1/stashes/file/{slug}/unlock": {"post": {"summary": "Unlock Protected File Stash", "operationId": "unlock_protected_file_stash_api_v1_stashes_file__slug__unlock_post", "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string", "title": "Slug"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "string", "writeOnly": true, "format": "password", "title": "Password"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/octet-stream": {}}}, "500": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Internal Server Error"}, "404": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Not Found"}, "400": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Bad Request"}, "410": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Gone"}, "401": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Unauthorized"}, "403": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Forbidden"}, "429": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Too Many Requests"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/stashes/text/{slug}/unlock": {"post": {"summary": "Unlock Protected Text Stash", "operationId": "unlock_protected_text_stash_api_v1_stashes_text__slug__unlock_post", "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string", "title": "Slug"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "string", "writeOnly": true, "format": "password", "title": "Password"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "string", "title": "Response Unlock Protected Text Stash Api V1 Stashes Text  Slug  Unlock Post"}}, "application/octet-stream": {}}}, "500": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Internal Server Error"}, "404": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Not Found"}, "400": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Bad Request"}, "410": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Gone"}, "401": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Unauthorized"}, "403": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Forbidden"}, "429": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Too Many Requests"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/stashes/file/{slug}": {"get": {"summary": "Get File Stash", "operationId": "get_file_stash_api_v1_stashes_file__slug__get", "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string", "title": "Slug"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/octet-stream": {}}}, "500": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Internal Server Error"}, "404": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Not Found"}, "400": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Bad Request"}, "410": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Gone"}, "401": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Unauthorized"}, "403": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Forbidden"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/stashes/text/{slug}": {"get": {"summary": "Get Text Stash", "operationId": "get_text_stash_api_v1_stashes_text__slug__get", "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string", "title": "Slug"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "string", "title": "Response Get Text Stash Api V1 Stashes Text  Slug  Get"}}}}, "500": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Internal Server Error"}, "404": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Not Found"}, "400": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Bad Request"}, "410": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Gone"}, "401": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Unauthorized"}, "403": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Forbidden"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/stashes/metadata/{slug}": {"get": {"summary": "Get Stash Metadata", "operationId": "get_stash_metadata_api_v1_stashes_metadata__slug__get", "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string", "title": "Slug"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GetStashBySlugRow"}}}}, "500": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Internal Server Error"}, "404": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Not Found"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/stashes/views/{slug}": {"get": {"summary": "Get Stash Views", "operationId": "get_stash_views_api_v1_stashes_views__slug__get", "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string", "title": "Slug"}}, {"name": "unique", "in": "query", "required": true, "schema": {"type": "boolean", "title": "Unique"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "integer", "title": "Response Get Stash Views Api V1 Stashes Views  Slug  Get"}}}}, "500": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Internal Server Error"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/google": {"get": {"summary": "Google Login", "operationId": "google_login_api_v1_auth_google_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GoogleLoginLocation"}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}, "401": {"description": "Unauthorized", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}}}, "/api/v1/auth/google/callback": {"get": {"summary": "Google Callback", "operationId": "google_callback_api_v1_auth_google_callback_get", "parameters": [{"name": "code", "in": "query", "required": true, "schema": {"type": "string", "title": "Code"}}, {"name": "state", "in": "query", "required": true, "schema": {"type": "string", "title": "State"}}], "responses": {"303": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "500": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Internal Server Error"}, "400": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Bad Request"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/google/me": {"get": {"summary": "Get Me", "operationId": "get_me_api_v1_auth_google_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"anyOf": [{"$ref": "#/components/schemas/User"}, {"type": "null"}], "title": "Response Get Me Api V1 Auth Google Me Get"}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}}}, "/api/v1/auth/google/logout": {"post": {"summary": "Google Logout", "operationId": "google_logout_api_v1_auth_google_logout_post", "responses": {"204": {"description": "Successful Response"}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}}}, "/api/v1/ip/{ip_addr}/activity": {"get": {"summary": "List Ip Activity", "operationId": "list_ip_activity_api_v1_ip__ip_addr__activity_get", "parameters": [{"name": "ip_addr", "in": "path", "required": true, "schema": {"type": "string", "title": "Ip Addr"}}, {"name": "page", "in": "query", "required": true, "schema": {"type": "integer", "exclusiveMinimum": 0, "title": "Page"}}, {"name": "take", "in": "query", "required": true, "schema": {"type": "integer", "exclusiveMaximum": 200, "exclusiveMinimum": 0, "title": "Take"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/ListIPActivityRow"}, "title": "Response List Ip Activity Api V1 Ip  Ip Addr  Activity Get"}}}}, "500": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Internal Server Error"}, "403": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Forbidden"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/ip/{ip_addr}/bans/active": {"get": {"summary": "Get Active Ip Ban", "operationId": "get_active_ip_ban_api_v1_ip__ip_addr__bans_active_get", "parameters": [{"name": "ip_addr", "in": "path", "required": true, "schema": {"type": "string", "title": "Ip Addr"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"anyOf": [{"$ref": "#/components/schemas/IpBan"}, {"type": "null"}], "title": "Response Get Active Ip Ban Api V1 Ip  Ip Addr  Bans Active Get"}}}}, "500": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Internal Server Error"}, "403": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Forbidden"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/ip/{ip_addr}/ban": {"post": {"summary": "Add Ip Ban", "operationId": "add_ip_ban_api_v1_ip__ip_addr__ban_post", "parameters": [{"name": "ip_addr", "in": "path", "required": true, "schema": {"type": "string", "title": "Ip Addr"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_add_ip_ban_api_v1_ip__ip_addr__ban_post"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IpBan"}}}}, "500": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Internal Server Error"}, "403": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Forbidden"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Unprocessable Content"}}}}, "/api/v1/ip/bans/{ban_id}": {"delete": {"summary": "Revoke Ban", "operationId": "revoke_ban_api_v1_ip_bans__ban_id__delete", "parameters": [{"name": "ban_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Ban Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Reason"}}}}, "responses": {"204": {"description": "Successful Response"}, "500": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Internal Server Error"}, "403": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "description": "Forbidden"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/config/max-upload-bytes": {"get": {"summary": "Get Config", "operationId": "get_config_api_v1_config_max_upload_bytes_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "integer", "title": "Response Get Config Api V1 Config Max Upload Bytes Get"}}}}, "500": {"description": "Internal Server Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}}}}}, "components": {"schemas": {"Body_add_binary_stash_api_v1_stashes_file_post": {"properties": {"file": {"type": "string", "contentMediaType": "application/octet-stream", "title": "File"}, "password": {"anyOf": [{"type": "string", "format": "password", "writeOnly": true}, {"type": "null"}], "title": "Password"}}, "type": "object", "required": ["file"], "title": "Body_add_binary_stash_api_v1_stashes_file_post"}, "Body_add_ip_ban_api_v1_ip__ip_addr__ban_post": {"properties": {"expires": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Expires"}, "reason": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Reason"}}, "type": "object", "required": ["expires", "reason"], "title": "Body_add_ip_ban_api_v1_ip__ip_addr__ban_post"}, "Body_add_text_stash_api_v1_stashes_text_post": {"properties": {"content": {"type": "string", "title": "Content"}, "password": {"anyOf": [{"type": "string", "format": "password", "writeOnly": true}, {"type": "null"}], "title": "Password"}}, "type": "object", "required": ["content"], "title": "Body_add_text_stash_api_v1_stashes_text_post"}, "GetStashBySlugRow": {"properties": {"id_": {"type": "integer", "title": "Id"}, "is_binary": {"type": "boolean", "title": "Is Binary"}, "slug": {"type": "string", "title": "Slug"}, "added": {"type": "string", "format": "date-time", "title": "Added"}, "added_by_ip": {"type": "string", "title": "Added By Ip"}, "added_by_user_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Added By User Id"}, "revoked_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Revoked At"}, "revoked_by_user_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Revoked By User Id"}, "expires_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Expires At"}, "is_protected": {"type": "boolean", "title": "Is Protected"}}, "type": "object", "required": ["id_", "is_binary", "slug", "added", "added_by_ip", "added_by_user_id", "revoked_at", "revoked_by_user_id", "expires_at", "is_protected"], "title": "GetStashBySlugRow"}, "GoogleLoginLocation": {"properties": {"url": {"type": "string", "title": "Url"}}, "type": "object", "required": ["url"], "title": "GoogleLoginLocation"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "IpBan": {"properties": {"id_": {"type": "integer", "title": "Id"}, "ip_address": {"type": "string", "title": "Ip Address"}, "added": {"type": "string", "format": "date-time", "title": "Added"}, "expires": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Expires"}, "reason": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Reason"}, "added_by_user_id": {"type": "integer", "title": "Added By User Id"}, "revoked_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Revoked At"}, "revoked_by_user_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Revoked By User Id"}, "revocation_reason": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Revocation Reason"}}, "type": "object", "required": ["id_", "ip_address", "added", "expires", "reason", "added_by_user_id", "revoked_at", "revoked_by_user_id", "revocation_reason"], "title": "IpBan"}, "ListIPActivityRow": {"properties": {"event_at": {"type": "string", "format": "date-time", "title": "Event At"}, "event_type": {"type": "string", "title": "Event Type"}, "stash_id": {"type": "integer", "title": "Stash Id"}, "slug": {"type": "string", "title": "Slug"}, "details": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Details"}}, "type": "object", "required": ["event_at", "event_type", "stash_id", "slug", "details"], "title": "ListIPActivityRow"}, "ListStashesRow": {"properties": {"id_": {"type": "integer", "title": "Id"}, "is_binary": {"type": "boolean", "title": "Is Binary"}, "slug": {"type": "string", "title": "Slug"}, "added": {"type": "string", "format": "date-time", "title": "Added"}, "added_by_ip": {"type": "string", "title": "Added By Ip"}, "added_by_user_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Added By User Id"}, "revoked_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Revoked At"}, "revoked_by_user_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Revoked By User Id"}, "expires_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Expires At"}, "is_protected": {"type": "boolean", "title": "Is Protected"}}, "type": "object", "required": ["id_", "is_binary", "slug", "added", "added_by_ip", "added_by_user_id", "revoked_at", "revoked_by_user_id", "expires_at", "is_protected"], "title": "ListStashesRow"}, "Message": {"properties": {"detail": {"type": "string", "title": "Detail"}}, "type": "object", "required": ["detail"], "title": "Message"}, "Stash": {"properties": {"id_": {"type": "integer", "title": "Id"}, "is_binary": {"type": "boolean", "title": "Is Binary"}, "slug": {"type": "string", "title": "Slug"}, "added": {"type": "string", "format": "date-time", "title": "Added"}, "added_by_ip": {"type": "string", "title": "Added By Ip"}, "added_by_user_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Added By User Id"}}, "type": "object", "required": ["id_", "is_binary", "slug", "added", "added_by_ip", "added_by_user_id"], "title": "Stash"}, "User": {"properties": {"id_": {"type": "integer", "title": "Id"}, "google_sub": {"type": "string", "title": "Google Sub"}, "email": {"type": "string", "title": "Email"}, "name": {"type": "string", "title": "Name"}, "created": {"type": "string", "format": "date-time", "title": "Created"}, "last_login": {"type": "string", "format": "date-time", "title": "Last Login"}, "is_admin": {"type": "boolean", "title": "Is Admin"}}, "type": "object", "required": ["id_", "google_sub", "email", "name", "created", "last_login", "is_admin"], "title": "User"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}